+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 0.10.0.0 - August 7 2014
+* Handle option types
+* Handle types with no properties/fields (oops)
+
 #### 0.9.0.0 - July 30 2014
 * Initial release